feat: enhance Vite app integration and update documentation
- Added `node-env-resolver` as a dependency in the Vite app example for improved environment variable handling.
- Updated README to clarify usage context for server and client variables, emphasizing the importance of using `import.meta.env` in browser code.
- Enhanced the Vite example with detailed explanations of the "Vite Way" and "Enhanced Way" for managing environment variables.
- Improved type safety and validation features in the Vite app configuration.
- Refactored example code to align with the latest API changes and best practices.
