+++ examples/basic/src/aws-without-zod-simple.test.ts
@@ -4,7 +4,8 @@
  */
 import { describe, it, expect } from 'vitest';
 import {  resolveAsync } from 'node-env-resolver';
-import { processEnv, string, url } from 'node-env-resolver/resolvers';
+import { processEnv } from 'node-env-resolver/resolvers';
+import { string, url } from 'node-env-resolver/validators';
 import { cached, TTL } from 'node-env-resolver/utils';
 import type { Resolver } from 'node-env-resolver';
