+++ packages/node-env-resolver/src/performance.test.ts
@@ -6,7 +6,7 @@
 
 import { describe, it, expect } from 'vitest';
 import { resolveAsync } from './index';
-import { string, url } from './resolvers';
+import { string, url } from './validators';
 import type { Resolver } from './types';
 
 // Helper to create mock resolvers with load tracking
