+++ packages/node-env-resolver/src/process-env.ts
@@ -3,6 +3,16 @@ import { SyncResolver } from "./types";
 /**
  * Resolver that reads from process.env
  * This is the default resolver used when no custom resolvers are provided
+ *
+ * @returns SyncResolver that loads environment variables from process.env
+ *
+ * @example
+ * ```ts
+ * import { processEnv } from 'node-env-resolver';
+ *
+ * const resolver = processEnv();
+ * const env = resolver.loadSync(); // { PORT: '3000', NODE_ENV: 'development', ... }
+ * ```
  */
 export function processEnv(): SyncResolver {
   return {
