+++ packages/node-env-resolver/src/resolvers.ts
@@ -4,12 +4,17 @@
 
 import { readFileSync, existsSync } from 'fs';
 import { join, resolve as resolvePath } from 'path';
-import type { Resolver, Validator } from './types';
+import type { Resolver } from './types';
 
 export { processEnv } from './process-env';
 
+/**
+ * Options for the dotenv resolver
+ */
 export interface DotenvOptions {
+  /** Path to the .env file (default: '.env') */
   path?: string;
+  /** Enable environment-specific file loading (default: false) */
   expand?: boolean;
 }
 
@@ -70,6 +75,31 @@ const parseDotenv = (content: string) => {
   return env;
 };
 
+/**
+ * Load configuration from .env files
+ * Supports environment-specific files when expand=true
+ *
+ * @param options Path string or options object
+ * @returns Resolver that loads from .env file(s)
+ *
+ * @example
+ * ```ts
+ * import { dotenv } from 'node-env-resolver/resolvers';
+ *
+ * // Simple .env file
+ * const config = await resolveAsync([
+ *   dotenv('.env'),
+ *   { PORT: 3000, API_KEY: string() }
+ * ]);
+ *
+ * // Environment-specific files
+ * const config = await resolveAsync([
+ *   dotenv({ path: '.env', expand: true }),
+ *   { PORT: 3000 }
+ * ]);
+ * // Loads: .env.defaults, .env, .env.local, .env.development, .env.development.local
+ * ```
+ */
 export function dotenv(options?: string | DotenvOptions): Resolver {
   const opts =
     typeof options === 'string'
@@ -504,6 +534,7 @@ export function secrets(path: string = '/run/secrets'): Resolver {
  * Requires 'js-yaml' as a peer dependency (install separately for tree-shaking)
  */
 export interface YamlOptions {
+  /** Path to the YAML file (default: 'config.yaml') */
   path?: string;
 }
 
@@ -630,6 +661,7 @@ export function yaml(options?: string | YamlOptions): Resolver {
  * Requires 'smol-toml' as a peer dependency (install separately for tree-shaking)
  */
 export interface TomlOptions {
+  /** Path to the TOML file (default: 'config.toml') */
   path?: string;
 }
 
@@ -749,693 +781,4 @@ export function toml(options?: string | TomlOptions): Resolver {
       }
     },
   };
-}
-
-export function string<
-  Opts extends {
-    default?: string;
-    optional?: boolean;
-    min?: number;
-    max?: number;
-    allowEmpty?: boolean;
-    pattern?: string;
-  } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & { optional: true; default?: Opts['default'] }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string, key?: string) => {
-    if (value === '' && !opts?.allowEmpty) {
-      throw new Error('String cannot be empty');
-    }
-    if (opts?.min !== undefined && value.length < opts.min) {
-      throw new Error(`String too short (min: ${opts.min})`);
-    }
-    if (opts?.max !== undefined && value.length > opts.max) {
-      throw new Error(`String too long (max: ${opts.max})`);
-    }
-    if (opts?.pattern && !new RegExp(opts.pattern).test(value)) {
-      const keyPrefix = key
-        ? `${key} does not match required pattern`
-        : 'String does not match pattern';
-      throw new Error(keyPrefix);
-    }
-    return value;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function number<
-  Opts extends {
-    default?: number;
-    optional?: boolean;
-    min?: number;
-    max?: number;
-  } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<number> & { optional: true; default?: Opts['default'] }
-  : Validator<number> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    const num = Number(value);
-    if (isNaN(num)) {
-      throw new Error(`Invalid number: "${value}"`);
-    }
-    if (opts?.min !== undefined && num < opts.min) {
-      throw new Error(`Number too small (min: ${opts.min})`);
-    }
-    if (opts?.max !== undefined && num > opts.max) {
-      throw new Error(`Number too large (max: ${opts.max})`);
-    }
-    return num;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function boolean<
-  Opts extends { default?: boolean; optional?: boolean } = Record<
-    string,
-    never
-  >,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<boolean> & { optional: true; default?: Opts['default'] }
-  : Validator<boolean> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    const lowerValue = value.toLowerCase();
-    if (['true', '1', 'yes', 'on'].includes(lowerValue)) {
-      return true;
-    }
-    if (['false', '0', 'no', 'off', ''].includes(lowerValue)) {
-      return false;
-    }
-    throw new Error(`Invalid boolean: "${value}"`);
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-// Additional validator functions
-export function url<
-  Opts extends { default?: string; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & { optional: true; default?: Opts['default'] }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    try {
-      new URL(value);
-      return value;
-    } catch {
-      throw new Error(`Invalid URL: "${value}"`);
-    }
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function email<
-  Opts extends { default?: string; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & { optional: true; default?: Opts['default'] }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    if (!emailRegex.test(value)) {
-      throw new Error(`Invalid email: "${value}"`);
-    }
-    return value;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function port<
-  Opts extends { default?: number; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<number> & { optional: true; default?: Opts['default'] }
-  : Validator<number> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    const num = Number(value);
-    if (isNaN(num) || num < 1 || num > 65535) {
-      throw new Error(`Invalid port: "${value}"`);
-    }
-    return num;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function postgres<
-  Opts extends { default?: string; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & { optional: true; default?: Opts['default'] }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    if (
-      !value.startsWith('postgres://') &&
-      !value.startsWith('postgresql://')
-    ) {
-      throw new Error(`Invalid PostgreSQL URL: "${value}"`);
-    }
-    return value;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function mysql<
-  Opts extends { default?: string; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & { optional: true; default?: Opts['default'] }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    if (!value.startsWith('mysql://')) {
-      throw new Error(`Invalid MySQL URL: "${value}"`);
-    }
-    return value;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function mongodb<
-  Opts extends { default?: string; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & { optional: true; default?: Opts['default'] }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    if (
-      !value.startsWith('mongodb://') &&
-      !value.startsWith('mongodb+srv://')
-    ) {
-      throw new Error(`Invalid MongoDB URL: "${value}"`);
-    }
-    return value;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function redis<
-  Opts extends { default?: string; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & { optional: true; default?: Opts['default'] }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    if (!value.startsWith('redis://') && !value.startsWith('rediss://')) {
-      throw new Error(`Invalid Redis URL: "${value}"`);
-    }
-    return value;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-
-export function https<
-  Opts extends { default?: string; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & { optional: true; default?: Opts['default'] }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    if (!value.startsWith('https://')) {
-      throw new Error(`Invalid HTTPS URL: "${value}"`);
-    }
-    return value;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function stringArray<
-  Opts extends {
-    default?: string[];
-    optional?: boolean;
-    separator?: string;
-  } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string[]> & { optional: true; default?: Opts['default'] }
-  : Validator<string[]> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    const separator = opts?.separator || ',';
-    return value
-      .split(separator)
-      .map((s) => s.trim())
-      .filter((s) => s.length > 0);
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function numberArray<
-  Opts extends {
-    default?: number[];
-    optional?: boolean;
-    separator?: string;
-  } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<number[]> & { optional: true; default?: Opts['default'] }
-  : Validator<number[]> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    const separator = opts?.separator || ',';
-    return value.split(separator).map((s) => {
-      const num = Number(s.trim());
-      if (isNaN(num)) {
-        throw new Error(`Invalid number in array: "${s}"`);
-      }
-      return num;
-    });
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function urlArray<
-  Opts extends {
-    default?: string[];
-    optional?: boolean;
-    separator?: string;
-  } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string[]> & { optional: true; default?: Opts['default'] }
-  : Validator<string[]> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    const separator = opts?.separator || ',';
-    const urls = value
-      .split(separator)
-      .map((s) => s.trim())
-      .filter((s) => s.length > 0);
-    for (const url of urls) {
-      try {
-        new URL(url);
-      } catch {
-        throw new Error(`Invalid URL in array: "${url}"`);
-      }
-    }
-    return urls;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function enums<
-  T extends readonly (string | number)[],
-  Opts extends { default?: T[number]; optional?: boolean } = Record<
-    string,
-    never
-  >,
->(
-  values: T,
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<T[number]> & { optional: true; default?: Opts['default'] }
-  : Validator<T[number]> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    if (!values.includes(value as T[number])) {
-      throw new Error(
-        `Invalid enum value: "${value}". Allowed values: ${values.join(', ')}`,
-      );
-    }
-    return value as T[number];
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function duration<
-  Opts extends { default?: number; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<number> & { optional: true; default?: Opts['default'] }
-  : Validator<number> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    // Simple duration parser - converts to milliseconds
-    const match = value.match(/^(\d+)([smhd])$/);
-    if (!match) {
-      throw new Error(
-        `Invalid duration: "${value}". Use format like "5s", "2m", "1h", "1d"`,
-      );
-    }
-    const [, num, unit] = match;
-    const multipliers = { s: 1000, m: 60000, h: 3600000, d: 86400000 };
-    const multiplier = multipliers[unit as keyof typeof multipliers];
-    return Number(num) * multiplier;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function file<
-  Opts extends {
-    default?: string;
-    optional?: boolean;
-    secretsDir?: string;
-  } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & {
-      optional: true;
-      default?: Opts['default'];
-      secretsDir?: Opts['secretsDir'];
-    }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-      secretsDir?: Opts['secretsDir'];
-    } {
-  const validator = ((value: string) => {
-    // Read file content from the provided path
-    try {
-      // eslint-disable-next-line @typescript-eslint/no-require-imports
-      const fs = require('fs');
-      // eslint-disable-next-line @typescript-eslint/no-require-imports
-      const { resolve } = require('path');
-      // Resolve file path relative to current working directory
-      const resolvedPath = resolve(process.cwd(), value);
-      const content = fs.readFileSync(resolvedPath, 'utf8').trim();
-      return content;
-    } catch (error) {
-      throw new Error(
-        `Failed to read file: ${error instanceof Error ? error.message : String(error)}`,
-      );
-    }
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-    (validator as Record<string, unknown>).secretsDir = opts.secretsDir;
-  }
-
-  // Mark this as a file validator for secretsDir support
-  (validator as Record<string, unknown>).__isFileValidator = true;
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function secret<
-  Opts extends { default?: string; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & { optional: true; default?: Opts['default'] }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    // For now, just return the secret value
-    return value;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function custom<
-  T = unknown,
-  Opts extends { default?: T; optional?: boolean } = Record<string, never>,
->(
-  validator: (value: string) => T,
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<T> & { optional: true; default?: Opts['default'] }
-  : Validator<T> & { optional?: Opts['optional']; default?: Opts['default'] } {
-  const validatorWithOptions = validator as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validatorWithOptions as Record<string, unknown>).default = opts.default;
-    (validatorWithOptions as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validatorWithOptions as any;
-}
-
-// Date and timestamp validators
-export function date<
-  Opts extends { default?: string; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<string> & { optional: true; default?: Opts['default'] }
-  : Validator<string> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    // Validate ISO 8601 date format - must match pattern like YYYY-MM-DD or YYYY-MM-DDTHH:mm:ssZ
-    const iso8601Pattern =
-      /^\d{4}-\d{2}-\d{2}(T\d{2}:\d{2}:\d{2}(\.\d{3})?Z?)?$/;
-    if (!iso8601Pattern.test(value)) {
-      throw new Error(`Date must be in ISO 8601 format: "${value}"`);
-    }
-    const date = new Date(value);
-    if (isNaN(date.getTime())) {
-      throw new Error(`Cannot parse date value: "${value}"`);
-    }
-    return value;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
-}
-
-export function timestamp<
-  Opts extends { default?: number; optional?: boolean } = Record<string, never>,
->(
-  opts?: Opts,
-): Opts extends { optional: true }
-  ? Validator<number> & { optional: true; default?: Opts['default'] }
-  : Validator<number> & {
-      optional?: Opts['optional'];
-      default?: Opts['default'];
-    } {
-  const validator = ((value: string) => {
-    const num = Number(value);
-    if (isNaN(num) || !Number.isInteger(num)) {
-      throw new Error(`Invalid timestamp: "${value}"`);
-    }
-    if (num < 0) {
-      throw new Error(`Invalid timestamp: "${value}"`);
-    }
-    if (num > 253402300799) {
-      // Year 9999
-      throw new Error(`Timestamp too large: "${value}"`);
-    }
-    return num;
-  }) as unknown;
-
-  // Attach options to the validator function for runtime
-  if (opts) {
-    (validator as Record<string, unknown>).default = opts.default;
-    (validator as Record<string, unknown>).optional = opts.optional;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return validator as any;
 }
